separate the folding rule from component file
